+++ dragonite-forwarder/src/main/java/com/vecsight/dragonite/forwarder/config/ForwarderServerConfig.java
@@ -60,6 +60,14 @@ public class ForwarderServerConfig {
         dragoniteSocketParameters.setEnableWebPanel(enabled);
     }
 
+    public InetSocketAddress getWebPanelBind() {
+        return dragoniteSocketParameters.getWebPanelBindAddress();
+    }
+
+    public void setWebPanelBind(InetSocketAddress address) {
+        dragoniteSocketParameters.setWebPanelBindAddress(address);
+    }
+
     public DragoniteSocketParameters getDragoniteSocketParameters() {
         return dragoniteSocketParameters;
     }